net/avf: fix iterator for RSS LUT
[ upstream commit 6faf884136f302b3bf0d2f984090b6c3816c7062 ]

Change RSS LUT iterator from uint8_t to uint16_t since the
RSS LUT size could exceed 255.

Fixes: 69dd4c3d0898 ("net/avf: enable queue and device")

Acked-by: Ting Xu <ting.xu@intel.com>

+++ drivers/net/avf/avf_ethdev.c
@@ -125,7 +125,7 @@ avf_init_rss(struct avf_adapter *adapter)
 {
 	struct avf_info *vf =  AVF_DEV_PRIVATE_TO_VF(adapter);
 	struct rte_eth_rss_conf *rss_conf;
-	uint8_t i, j, nb_q;
+	uint16_t i, j, nb_q;
 	int ret;
 
 	rss_conf = &adapter->eth_dev->data->dev_conf.rx_adv_conf.rss_conf;
